Refactor: professions list as new function.
