+++ src/views/components/searchBarCard.tsx
