+++ src/views/pages/customerHome.tsx
@@ -1,33 +1,12 @@
 import { Box, ButtonBase, Typography } from '@mui/material'
 import { useState } from 'react'
-import ConstructionIcon from '@mui/icons-material/Construction'
-import BuildIcon from '@mui/icons-material/Build'
-import PlumbingIcon from '@mui/icons-material/Plumbing'
-import CarpenterIcon from '@mui/icons-material/Carpenter'
-import DesignServicesIcon from '@mui/icons-material/DesignServices'
-import FormatPaintIcon from '@mui/icons-material/FormatPaint'
-import ElectricalServicesIcon from '@mui/icons-material/ElectricalServices'
-import PsychologyAltIcon from '@mui/icons-material/PsychologyAlt'
-import SearchBarCard from '../components/containers/searchBarCard'
+import SearchBarCard from '../components/searchBarCard'
+import Professions from '../../models/professions'
 
 export default function CustomerHome() {
-  const iconList = [
-    { icon: <ConstructionIcon sx={styles.icon} />, label: 'Albañil' },
-    { icon: <BuildIcon sx={styles.icon} />, label: 'Mecanico' },
-    { icon: <PlumbingIcon sx={styles.icon} />, label: 'Plomero' },
-    { icon: <CarpenterIcon sx={styles.icon} />, label: 'Carpintero' },
-    { icon: <DesignServicesIcon sx={styles.icon} />, label: 'Arquitecto' },
-    { icon: <FormatPaintIcon sx={styles.icon} />, label: 'Pintor' },
-    {
-      icon: <ElectricalServicesIcon sx={styles.icon} />,
-      label: 'Electricista',
-    },
-    { icon: <PsychologyAltIcon sx={styles.icon} />, label: 'Otro' },
-  ]
-
   const [searchTerm, setSearchTerm] = useState('')
 
-  const filteredIconList = iconList.filter((item) =>
+  const filteredIconList = Professions().filter((item) =>
     item.label.toLowerCase().includes(searchTerm.toLowerCase()),
   )
 
@@ -80,10 +59,5 @@ const styles = {
     width: '100px',
     height: '100px',
     padding: '10px',
-  },
-
-  icon: {
-    width: '50px',
-    height: '60px',
   },
 }
